third_party/blink/web_tests/webaudio/unit-tests/jstest-test-expected.txt

A simple unit testing for audio-testing.js and js-test.js

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Zero is equal to 0.
PASS One is not equal to 0.
PASS Expected SNR (110) is greater than or equal to 100.
PASS Maximum error value (0.000001) is less than or equal to 0.00001.
FAIL max error (0.000001) is not less than or equal to -1.
PASS One point double zero one is 1 within a relative error of 0.1.
FAIL Two is not 1 within a relative error of 0.1: 2 with relative error 1.
PASS [2, 2, 2] contains only the constant 2.
PASS [1, 2, 3] is identical to the array [1,2,3].
PASS My array equals [0.1,0.2] with an element-wise tolerance of 0.02.
PASS My random array contains all the expected values in the correct order: [1,3,2].
PASS successfullyParsed is true

TEST COMPLETE
